+++ source/common/stats/BUILD
@@ -75,4 +75,9 @@ envoy_cc_library(
     name = "hystrix_lib",
     srcs = ["hystrix.cc"],
     hdrs = ["hystrix.h"],
+    deps = [
+        "//include/envoy/server:admin_interface",
+        "//include/envoy/server:instance_interface",
+        "//source/common/buffer:buffer_lib",
+    ],
 )
