+++ test/server/http/admin_test.cc
@@ -99,7 +99,7 @@ INSTANTIATE_TEST_CASE_P(IpVersions, AdminInstanceTest,
 TEST_P(AdminInstanceTest, AdminProfiler) {
   Buffer::OwnedImpl data;
   Http::HeaderMapImpl header_map;
-  HandlerInfo handler_info;
+  HandlerInfoImpl handler_info;
   EXPECT_EQ(Http::Code::OK,
             admin_.runCallback("/cpuprofiler?enable=y", header_map, data, handler_info));
   EXPECT_TRUE(Profiler::Cpu::profilerEnabled());
@@ -117,7 +117,7 @@ TEST_P(AdminInstanceTest, AdminBadProfiler) {
                                    "", Network::Test::getCanonicalLoopbackAddress(GetParam()),
                                    server_, listener_scope_.createScope("listener.admin."));
   Http::HeaderMapImpl header_map;
-  HandlerInfo handler_info;
+  HandlerInfoImpl handler_info;
   admin_bad_profile_path.runCallback("/cpuprofiler?enable=y", header_map, data, handler_info);
   EXPECT_FALSE(Profiler::Cpu::profilerEnabled());
 }
@@ -145,7 +145,7 @@ TEST_P(AdminInstanceTest, CustomHandler) {
   EXPECT_TRUE(admin_.addHandler("/foo/bar", "hello", callback, true, false));
   Http::HeaderMapImpl header_map;
   Buffer::OwnedImpl response;
-  HandlerInfo handler_info;
+  HandlerInfoImpl handler_info;
 
   EXPECT_EQ(Http::Code::Accepted,
             admin_.runCallback("/foo/bar", header_map, response, handler_info));
@@ -187,7 +187,7 @@ TEST_P(AdminInstanceTest, EscapeHelpTextWithPunctuation) {
   auto callback = [&](const std::string&, Http::HeaderMap&, Buffer::Instance&,
                       HandlerInfo&) -> Http::Code { return Http::Code::Accepted; };
 
-  HandlerInfo handler_info;
+  HandlerInfoImpl handler_info;
 
   // It's OK to have help text with HTML characters in it, but when we render the home
   // page they need to be escaped.
@@ -207,7 +207,7 @@ TEST_P(AdminInstanceTest, EscapeHelpTextWithPunctuation) {
 TEST_P(AdminInstanceTest, HelpUsesFormForMutations) {
   Http::HeaderMapImpl header_map;
   Buffer::OwnedImpl response;
-  HandlerInfo handler_info;
+  HandlerInfoImpl handler_info;
   EXPECT_EQ(Http::Code::OK, admin_.runCallback("/", header_map, response, handler_info));
   const std::string logging_action = "<form action='/logging' method='post'";
   const std::string stats_href = "<a href='/stats'";
@@ -223,7 +223,7 @@ TEST_P(AdminInstanceTest, Runtime) {
       {"string_key", {"foo", {}}}, {"int_key", {"1", {1}}}, {"other_key", {"bar", {}}}};
   Runtime::MockSnapshot snapshot;
   Runtime::MockLoader loader;
-  HandlerInfo handler_info;
+  HandlerInfoImpl handler_info;
 
   EXPECT_CALL(snapshot, getAll()).WillRepeatedly(testing::ReturnRef(entries));
   EXPECT_CALL(loader, snapshot()).WillRepeatedly(testing::ReturnPointee(&snapshot));
@@ -241,7 +241,7 @@ TEST_P(AdminInstanceTest, RuntimeJSON) {
       {"string_key", {"foo", {}}}, {"int_key", {"1", {1}}}, {"other_key", {"bar", {}}}};
   Runtime::MockSnapshot snapshot;
   Runtime::MockLoader loader;
-  HandlerInfo handler_info;
+  HandlerInfoImpl handler_info;
 
   EXPECT_CALL(snapshot, getAll()).WillRepeatedly(testing::ReturnRef(entries));
   EXPECT_CALL(loader, snapshot()).WillRepeatedly(testing::ReturnPointee(&snapshot));
@@ -277,7 +277,7 @@ TEST_P(AdminInstanceTest, RuntimeBadFormat) {
   std::unordered_map<std::string, const Runtime::Snapshot::Entry> entries;
   Runtime::MockSnapshot snapshot;
   Runtime::MockLoader loader;
-  HandlerInfo handler_info;
+  HandlerInfoImpl handler_info;
 
   EXPECT_CALL(snapshot, getAll()).WillRepeatedly(testing::ReturnRef(entries));
   EXPECT_CALL(loader, snapshot()).WillRepeatedly(testing::ReturnPointee(&snapshot));
